requirements.txt
# Core
streamlit==1.35.0
pandas==2.2.2
numpy==1.26.4
scipy==1.13.0

# Visualization
plotly==5.22.0
altair==5.3.0
matplotlib==3.9.0
seaborn==0.13.2

# ML
scikit-learn==1.5.0
xgboost==2.0.3
lightgbm==4.3.0
imbalanced-learn==0.12.3
shap==0.45.1

# NLP
transformers==4.41.1
sentence-transformers==3.0.1
torch==2.3.0
faiss-cpu==1.8.0
nltk==3.8.1
textblob==0.18.0

# Utilities
joblib==1.4.2
tqdm==4.66.4
python-dotenv==1.0.1
openpyxl==3.1.3